fix : ChannelCreateRequest @NotBlank 적용, ChannelController @Valid 적용, ChannelServiceImpl 삭제, 버전별 패키지 분리
