+++ src/main/java/com/fitian/burntz/domain/channel/v1/dto/ChannelCreateRequest.java
@@ -1,6 +1,9 @@
-package com.fitian.burntz.domain.channel.dto;
+package com.fitian.burntz.domain.channel.v1.dto;
 
 import com.fitian.burntz.domain.channel.enums.ChannelType;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -21,9 +24,18 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ChannelCreateRequest {
+    @NotBlank(message = "boxCode must not be blank")
     private String boxCode;
+
+    @NotBlank(message = "channelId must not be blank")
     private String channelId;
+
+    @NotBlank(message = "channelName must not be blank")
     private String channelName;
+
+    @NotNull(message = "type is required")
     private ChannelType type;
+
+    @NotEmpty(message = "memberPks must contain at least one member")
     private List<Long> memberPks;
 }
